main: split redis connection from redis cache

+++ config/database.php
@@ -8,7 +8,7 @@ use Tuto\Database\Pdo\PdoConnection;
 return [
     'database.type' => env('DB_TYPE', 'mysql'),
     'database.host' => env('DB_HOST', 'localhost'),
-    'database.port' => env('DB_PORT', 3306),
+    'database.port' => (int) env('DB_PORT', 3306),
     'database.username' => env('DB_USERNAME', 'root'),
     'database.password' => env('DB_PASSWORD', 'root'),
     'database.name' => env('DB_NAME', 'project'),
